refactor(settings): split /settings/organization into /settings/appearance and /settings/auth
Appearance keeps logos/favicon/theme/branding/site-notifications;
the new Auth page takes OAuth token lifetime, require-2FA, session
lifetime, default role, and the organization token card. Save-bar
consistency: Site Notifications loses its own Save button and joins the
appearance page's single floating save bar, and the Auth page is ONE
react-hook-form with ONE floating save bar issuing a single PATCH of the
dirty subset (previously three separate per-section save bars).
/settings/organization 308-redirects to /settings/appearance; tabs,
sidebar, page-permission map, roles-page copy, and the manage-token
deep-link all updated.
