+++ platform/frontend/src/app/settings/appearance/_components/chat-links-editor.utils.test.ts
