+++ platform/frontend/src/app/settings/appearance/_components/chat-placeholders-editor.tsx
