+++ platform/frontend/src/app/settings/appearance/_components/favicon-upload.tsx
