+++ platform/frontend/src/app/settings/appearance/_components/image-upload.tsx
