+++ platform/frontend/src/app/settings/appearance/_components/logos-section.tsx
