+++ platform/frontend/src/app/settings/appearance/_components/onboarding-wizards-editor.utils.test.ts
