+++ platform/frontend/src/app/settings/auth/_components/organization-token-section.tsx
