+++ platform/frontend/src/app/settings/auth/page.test.tsx
@@ -23,17 +23,56 @@ let mockOrganizationPending = false;
 
 vi.mock("@/lib/organization.query");
 vi.mock("@/lib/auth/auth.query");
+vi.mock("@/lib/config/config.query");
+vi.mock("@/lib/hooks/use-app-name");
+vi.mock("next/navigation");
 
+vi.mock("@/lib/role.query", () => ({
+  useRoles: vi.fn(() => ({
+    data: [
+      {
+        id: "role-member",
+        role: "member",
+        name: "member",
+        predefined: true,
+        permission: {},
+      },
+      {
+        id: "role-admin",
+        role: "admin",
+        name: "admin",
+        predefined: true,
+        permission: {},
+      },
+    ],
+    isPending: false,
+  })),
+}));
+
+vi.mock("@/lib/teams/team-token.query", () => ({
+  useTokens: vi.fn(() => ({
+    data: { tokens: [] },
+    isLoading: false,
+    error: null,
+  })),
+}));
+
+import { useSearchParams } from "next/navigation";
 import {
+  useAllPermissions,
   useHasPermissions,
   useMissingPermissions,
 } from "@/lib/auth/auth.query";
 import {
-  useAppearanceSettings,
+  useEnterpriseFeature,
+  useSmallTeamTier,
+} from "@/lib/config/config.query";
+import { useAppName } from "@/lib/hooks/use-app-name";
+import {
   useOrganization,
   useUpdateAuthSettings,
 } from "@/lib/organization.query";
-import { OAuthTokenLifetimeSection } from "./oauth-token-lifetime-section";
+import AuthSettingsPage from "./page";
 
 function renderPage() {
   const queryClient = new QueryClient({
@@ -42,21 +81,25 @@ function renderPage() {
 
   return render(
     <QueryClientProvider client={queryClient}>
-      <OAuthTokenLifetimeSection />
+      <AuthSettingsPage />
     </QueryClientProvider>,
   );
 }
 
-describe("OAuthTokenLifetimeSection", () => {
+describe("AuthSettingsPage", () => {
   beforeEach(() => {
     vi.clearAllMocks();
     mockOrganization = {
       oauthAccessTokenLifetimeSeconds: 31_536_000,
+      sessionMaxAgeSeconds: null,
+      requireTwoFactor: false,
+      defaultMemberRole: null,
     };
     mockOrganizationPending = false;
-    mutateAsync.mockResolvedValue({
-      oauthAccessTokenLifetimeSeconds: 604_800,
-    });
+    mutateAsync.mockImplementation(async (body: Record<string, unknown>) => ({
+      ...mockOrganization,
+      ...body,
+    }));
 
     vi.mocked(useOrganization).mockImplementation(
       () =>
@@ -65,13 +108,18 @@ describe("OAuthTokenLifetimeSection", () => {
           isPending: mockOrganizationPending,
         }) as ReturnType<typeof useOrganization>,
     );
-    vi.mocked(useAppearanceSettings).mockReturnValue({
-      data: { appName: null },
-    } as unknown as ReturnType<typeof useAppearanceSettings>);
     vi.mocked(useUpdateAuthSettings).mockReturnValue({
       mutateAsync,
       isPending: false,
     } as unknown as ReturnType<typeof useUpdateAuthSettings>);
+    vi.mocked(useEnterpriseFeature).mockReturnValue(true);
+    vi.mocked(useSmallTeamTier).mockReturnValue(
+      undefined as ReturnType<typeof useSmallTeamTier>,
+    );
+    vi.mocked(useAppName).mockReturnValue("Archestra");
+    vi.mocked(useSearchParams).mockReturnValue(
+      new URLSearchParams() as unknown as ReturnType<typeof useSearchParams>,
+    );
     vi.mocked(useHasPermissions).mockReturnValue({
       data: true,
       isPending: false,
@@ -79,6 +127,9 @@ describe("OAuthTokenLifetimeSection", () => {
     vi.mocked(useMissingPermissions).mockReturnValue(
       [] as unknown as ReturnType<typeof useMissingPermissions>,
     );
+    vi.mocked(useAllPermissions).mockReturnValue({
+      data: undefined,
+    } as ReturnType<typeof useAllPermissions>);
   });
 
   it("submits a preset OAuth token lifetime", async () => {
@@ -118,6 +169,75 @@ describe("OAuthTokenLifetimeSection", () => {
         oauthAccessTokenLifetimeSeconds: 123_456,
       });
     });
+  });
+
+  it("saves every dirty auth field in a single PATCH", async () => {
+    const user = userEvent.setup();
+
+    renderPage();
+
+    // OAuth token lifetime: 1 year -> 7 days
+    await user.click(screen.getByRole("combobox", { name: /token lifetime/i }));
+    await user.click(screen.getByRole("option", { name: "7 days" }));
+
+    // Require 2FA: off -> on
+    await user.click(screen.getByRole("switch"));
+
+    // Session lifetime: no limit -> 24 hours
+    await user.click(
+      screen.getByRole("combobox", { name: /maximum session lifetime/i }),
+    );
+    await user.click(screen.getByRole("option", { name: "24 hours" }));
+
+    // Default role: member -> admin
+    await user.click(screen.getByTestId("default-member-role-select"));
+    await user.click(screen.getByRole("option", { name: /admin/i }));
+
+    await user.click(screen.getByRole("button", { name: "Save" }));
+
+    await waitFor(() => {
+      expect(mutateAsync).toHaveBeenCalledTimes(1);
+      expect(mutateAsync).toHaveBeenCalledWith({
+        oauthAccessTokenLifetimeSeconds: 604_800,
+        sessionMaxAgeSeconds: 86_400,
+        requireTwoFactor: true,
+        defaultMemberRole: "admin",
+      });
+    });
+  });
+
+  it("discards changes without saving when cancelled", async () => {
+    const user = userEvent.setup();
+
+    renderPage();
+
+    await user.click(screen.getByRole("combobox", { name: /token lifetime/i }));
+    await user.click(screen.getByRole("option", { name: "7 days" }));
+
+    // Cancel is a native submit button inside the page form; the submit
+    // handler must diff the reset values and skip the PATCH entirely.
+    await user.click(screen.getByRole("button", { name: "Cancel" }));
+
+    await waitFor(() => {
+      expect(
+        screen.queryByRole("button", { name: "Save" }),
+      ).not.toBeInTheDocument();
+    });
+    expect(mutateAsync).not.toHaveBeenCalled();
+    expect(
+      screen.getByRole("combobox", { name: /token lifetime/i }),
+    ).toHaveTextContent("1 year");
+  });
+
+  it("hides the enterprise-only sections without a license", () => {
+    vi.mocked(useEnterpriseFeature).mockReturnValue(false);
+
+    renderPage();
+
+    expect(screen.queryByRole("switch")).not.toBeInTheDocument();
+    expect(
+      screen.queryByRole("combobox", { name: /maximum session lifetime/i }),
+    ).not.toBeInTheDocument();
   });
 
   it("shows the default preset when the organization response is missing the lifetime", () => {
